pyproject.toml
[project]
name = "cyber-poc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "camel-ai>=0.2.90",
    "fastapi>=0.135.3",
    "google-search-results>=2.4.2",
    "pydantic>=2.12.0",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "requests>=2.33.1",
    "sse-starlette>=3.3.4",
    "sseclient-py>=1.9.0",
    "uvicorn>=0.42.0",
]
